Fold the util/chplenv/chpl_3p_* scripts into the chpl_* ones.
We've had util/chplenv/chpl_3p_*_configs.py scripts for the various
third-party packages that defined the functions get_uniq_cfg_path(),
get_link_args(), get_jemalloc_config_file() (jemalloc only), and
get_compile_args() (libfabric only).

Here, fold those functions into other scripts that, except as noted,
already existed:
chpl_3p_gasnet_configs.py -> chpl_gasnet.py (didn't exist before)
chpl_3p_gmp_configs.py -> chpl_gmp.py
chpl_3p_hwloc_configs.py -> chpl_hwloc.py
chpl_3p_jemalloc_configs.py -> chpl_jemalloc.py
chpl_3p_libfabric_configs.py -> chpl_libfabric.py
chpl_3p_libunwind_configs.py -> chpl_unwind.py
chpl_3p_qthreads_configs.py -> chpl_qthreads.py (didn't exist before)
chpl_3p_re2_configs.py -> chpl_regexp.py
