--- util/chplenv/chpl_3p_re2_configs.py
@@ -1,12 +1,0 @@
-import third_party_utils
-from utils import memoize
-
-
-@memoize
-def get_uniq_cfg_path():
-    return third_party_utils.default_uniq_cfg_path()
-
-@memoize
-def get_link_args():
-    return third_party_utils.default_get_link_args('re2',
-                                                   libs=['-lre2', '-lpthread'])
