+++ util/test/re2_supports_valgrind.py
@@ -6,13 +6,13 @@ chplenv_dir = os.path.join(os.path.dirname(__file__), '..', 'chplenv')
 sys.path.insert(0, os.path.abspath(chplenv_dir))
 
 import chpl_home_utils
-import chpl_3p_re2_configs
+import chpl_regexp
 
 def get():
     """ Detects if re2 was built with CHPL_RE2_VALGRIND_SUPPORT set by checking
         for a sentinel file in the install dir """
     third_party = chpl_home_utils.get_chpl_third_party()
-    uniq_cfg_path = chpl_3p_re2_configs.get_uniq_cfg_path()
+    uniq_cfg_path = chpl_regexp.get_uniq_cfg_path()
     re2_valgrind = os.path.join(third_party, 're2', 'install',
                                 uniq_cfg_path, 'CHPL_RE2_VALGRIND_SUPPORT')
     return os.path.exists(re2_valgrind)
